Made the TeamPics Component Responsive!
I used a different source online that had code for a responsive hexagon design, and edited it. I also added media queries for the Header of the component, "Our Team", so the font-size changes based on a specified resolution's width. We also decided not to use the linkedIn photos because the default resolution on those are too small, and would need to be scaled. For now, there are placeholders because members need to upload their photos into our Google Drive.