(Media Query) max-width: 850px: Changed # of hexagons per row, & size
The size of the hexagon pictures were too small for width 850px and below, so I made a media query (max-width: 850px) that changed the amount of hexagons per row and made them bigger sized. Again, the centering of the hexagons were confusing, so I had to finish up with changing the div with className="CenterHexagons" 's margin-right.