Created a Sponsors Component for Designathon
The placeholders that are there currently show that if we test responsiveness by changing the window resolution, the placeholder logos would be hard to read. Thus an easy fix would be to upload sponsor logos that take up the entire width and height of the picture, so that each grid can be filled completely. Item1 needs to be replaced with the title asset.