Frontend Fixes for Designathon FAQ Component
When the browser width is lower than 500px, the FAQ titles and contents flex under each other. However because of this the width of the title and content containers would go off the screen, and thus allow the user to swipe left and right on their browser. Before these fixes, you would be able to go off-screen starting at width 500px and lower, and see a gray box for the contents off-screen. With these fixes, you would be able to go off-screen under width 333px. I had to adjust some height and width too in order to make sure the fixes worked on iPhone5/SE.