Added Shawn's pic to hexagons, cleaned up some CSS
Added Shawn's pic to hexagons. Saw that the HackMerced Tower was placed very low on the home page, so I changed the margins, height, and widths of it. Changed CSS and media queries so the tower, title img, title texts, and subscribe button have proper z-indexes when overlapping, and fixed minor responsiveness issues relating to that.