Changed Home Heading height w/ heading subtext & button margin
The Home Heading's img, text, subtext, and subscribe button were too squished together. I had to change the height of the Home Heading component and add css to that, and fix the css margins between the subtext and subscribe button. By doing this we fill most of the empty space in our Heading component; especially when at lower resolutions.