Heading Component's height is 100vh if res width 650+
Had to subtract the vh by the height of the navbar's height. Will not do vh for lower resolutions because lower resolutions are for phones and other devices. Apparently vh might not work on them and create visual bugs (because you can zoom-in and scroll out of the page using a phone, changing the view height).