Fixed Past-Sponsors-Logos Component code messing with Home Page imgs
Made the Sponsor-Us Page scss and HTML code more readable. Got rid of useless code. Changed 'id' to 'className'. In the scss file, '.img{}' and '#' were messing up the aspect ratio of images on the Home Page, and Sponsor Us Page. The mobile view 'X' button in the hamburger menu was out of view, but it is now visible again.