Changed Home Heading Component to have display: grid;
Grouped home heading's Title Image, Text, Subtext, and Button. Used display:grid to separate the tower from the grouped. These changes made the code cleaner and easier to read, and makes it so no CSS flexing of text in the home heading component will affect the CSS height of the tower.